# proxy/openapi/generated/recommand.yaml
openapi: 3.1.0
info:
  title: Recommand Peppol API
  description: >-

    Welcome to the Recommand Peppol API documentation. This API provides a
    comprehensive set of endpoints for managing and interacting with the
    Recommand Peppol platform.


    ## Getting Started


    To get started with the API:


    1. Create an API key and secret in the Recommand dashboard

    2. Use Basic Authentication with your API key and secret

    3. Make requests to the available endpoints


    ## Authentication


    All API requests must be authenticated using Basic Authentication. Your API
    key should be used as the username, and your API secret as the password.


    > [!TIP]

    > If you are signed in to the Recommand dashboard in this browser, you can
    use the API client that is embedded below without creating an API key, you
    will be authenticated via a session cookie.


    ## Team and company IDs


    For some endpoints, you will need to provide a team or company ID.


    - You can find your team ID [here](/api-keys).

    - You can find your company IDs [here](/companies).


    ## Support


    For additional support or questions, don't hesitate to contact our support
    team.
  version: 1.0.0
servers:
  - url: 'https://peppol.recommand.eu'
    description: Recommand API
  - url: 'https://test.ion-ap.net'
components:
  securitySchemes:
    httpBasic:
      type: http
      scheme: basic
      description: >-
        Basic API key authentication. Create a new API key and secret in the
        Recommand dashboard.
  schemas:
    Email:
      type: object
      properties:
        when:
          type: string
          enum:
            - always
            - on_peppol_failure
          default: on_peppol_failure
          description: When to send the email.
        to:
          type: array
          items:
            type: string
          description: The email addresses to send the document to.
          example:
            - support@recommand.eu
        subject:
          type: string
          description: >-
            The subject of the email. If not provided, the subject will be
            autogenerated based on the document type.
          example: Invoice SI-001
        htmlBody:
          type: string
          description: >-
            The HTML body of the email. If not provided, the body will be
            autogenerated based on the document type.
          example: 'Dear customer, you can find your invoice attached.'
      required:
        - to
      description: >-
        Send your document over email as well. You can choose to always send the
        email, or only when we cannot reach the recipient via Peppol. Each sent
        email is counted towards your document quota. When an email is expected
        to be sent, the request will succeed as long as an email has been sent,
        even if the Peppol sending failed (e.g. due to a non-existing Peppol
        address).
...
    Party:
      type: object
      properties:
        vatNumber:
          type:
            - string
            - 'null'
          example: BE0123456789
        name:
          type: string
          example: Example Company
        street:
          type: string
          example: Example Street 1
        street2:
          type:
            - string
            - 'null'
          example: Suite 100
        city:
          type: string
          example: Brussels
        postalZone:
          type: string
          example: '1000'
        country:
          type: string
          minLength: 2
          maxLength: 2
          example: BE
      required:
        - name
        - street
        - city
        - postalZone
        - country
    PaymentMeans:
      type: object
      properties:
        paymentMethod:
          type: string
          enum:
            - credit_transfer
          default: credit_transfer
          example: credit_transfer
        reference:
          type: string
          default: ''
          example: INV-2026-001
        iban:
          type: string
          example: BE1234567890
      required:
        - paymentMethod
        - reference
        - iban
    Line:
      type: object
      properties:
        name:
          type: string
          default: ''
          example: Consulting Services
        description:
          type:
            - string
            - 'null'
          example: Professional consulting services
        sellersId:
          type:
            - string
            - 'null'
          example: CS-001
        quantity:
          type: string
          example: '21.00'
          description: Decimal number as a string with flexible precision
          default: '1.00'
        unitCode:
          type: string
          default: C62
          example: HUR
          description: >-
            Recommended unit codes can be found
            [here](https://docs.peppol.eu/poacc/billing/3.0/codelist/UNECERec20/).
        netPriceAmount:
          type: string
          example: '21.00'
          description: Decimal number as a string with flexible precision
        netAmount:
          type:
            - string
            - 'null'
          example: '21.00'
          description: >-
            The total net amount of the line: quantity * netPriceAmount. Rounded
            to 2 decimal places. If not provided, it will be calculated
            automatically.
        vat:
          $ref: '#/components/schemas/VAT'
      required:
        - name
        - quantity
        - unitCode
        - netPriceAmount
        - vat
    VAT:
      type: object
      properties:
        category:
          type: string
          enum:
            - 'AE: Vat Reverse Charge'
            - 'E: Exempt from Tax'
            - 'S: Standard rate'
            - 'Z: Zero rated goods'
            - 'G: Free export item, VAT not charged'
            - 'O: Services outside scope of tax'
            - 'K: VAT exempt for EEA intra-community supply'
            - 'L: Canary Islands general indirect tax'
            - >-
              M: Tax for production, services and importation in Ceuta and
              Melilla
            - 'B: Transferred (VAT), In Italy'
          example: S
          description: >-
            VAT category code. When sending regular invoices, you should most
            often use the `S` category. When sending an invoice to another EU
            country, use the `AE` category for VAT Reverse Charge. In those
            cases, it is still recommended to include a note in the invoice
            explaining that the VAT Reverse Charge applies.
          default: S
        percentage:
          type: string
          example: '21.00'
          description: Decimal number as a string with 2 decimal places
      required:
        - category
        - percentage
    Totals:
      type: object
      properties:
        taxExclusiveAmount:
          type: string
          example: '21.00'
          description: Decimal number as a string with 2 decimal places
        taxInclusiveAmount:
          type: string
          example: '21.00'
          description: Decimal number as a string with 2 decimal places
        payableAmount:
          type:
            - string
            - 'null'
          example: '21.00'
          description: >-
            The amount to be paid. If not provided, this will be
            taxInclusiveAmount. Can be used in combination with paidAmount to
            indicate partial payment or payment rounding. Rounded to 2 decimal
            places.
        paidAmount:
          type:
            - string
            - 'null'
          example: '21.00'
          description: >-
            The amount paid. If not provided, this will be taxInclusiveAmount -
            payableAmount. Can be used in combination with payableAmount to
            indicate partial payment or payment rounding. Rounded to 2 decimal
            places.
      required:
        - taxExclusiveAmount
        - taxInclusiveAmount
      description: 'If not provided, the totals will be calculated from the document lines.'
    VatTotals:
      type: object
      properties:
        totalVatAmount:
          type: string
          example: '21.00'
          description: Decimal number as a string with 2 decimal places
        subtotals:
          type: array
          items:
            $ref: '#/components/schemas/VATSubtotal'
      required:
        - totalVatAmount
        - subtotals
      description: >-
        If not provided, the VAT totals will be calculated from the document
        lines.
    VATSubtotal:
      type: object
      properties:
        taxableAmount:
          type: string
          example: '21.00'
          description: Decimal number as a string with 2 decimal places
        vatAmount:
          type: string
          example: '21.00'
          description: Decimal number as a string with 2 decimal places
        category:
          type: string
          enum:
            - 'AE: Vat Reverse Charge'
            - 'E: Exempt from Tax'
            - 'S: Standard rate'
            - 'Z: Zero rated goods'
            - 'G: Free export item, VAT not charged'
            - 'O: Services outside scope of tax'
            - 'K: VAT exempt for EEA intra-community supply'
            - 'L: Canary Islands general indirect tax'
            - >-
              M: Tax for production, services and importation in Ceuta and
              Melilla
            - 'B: Transferred (VAT), In Italy'
          example: S
          description: >-
            VAT category code. When sending regular invoices, you should most
            often use the `S` category. When sending an invoice to another EU
            country, use the `AE` category for VAT Reverse Charge. In those
            cases, it is still recommended to include a note in the invoice
            explaining that the VAT Reverse Charge applies.
        percentage:
          type: string
          example: '21.00'
          description: Decimal number as a string with 2 decimal places
        exemptionReasonCode:
          type:
            - string
            - 'null'
          description: "If the invoice is exempt from VAT, this is required. The exemption reason code identifier must belong to the CEF VATEX code list\tfound [here](https://docs.peppol.eu/poacc/billing/3.0/2024-Q4/codelist/vatex/)."
      required:
        - taxableAmount
        - vatAmount
        - category
        - percentage
    Attachment:
      type: object
      properties:
        id:
          type: string
          example: ATT-001
        mimeCode:
          type: string
          default: application/pdf
          example: application/pdf
          description: >-
            MIME type of the document (e.g. application/pdf, text/csv,
            image/png)
        filename:
          type: string
          example: contract.pdf
        description:
          type:
            - string
            - 'null'
          example: Signed contract
        embeddedDocument:
          type:
            - string
            - 'null'
          description: base64 encoded document
        url:
          type:
            - string
            - 'null'
          example: 'https://example.com/contract.pdf'
      required:
        - id
        - mimeCode
        - filename
    SendInvoice:
      type: object
      properties:
        invoiceNumber:
          type: string
          example: INV-2024-001
        issueDate:
          type:
            - string
            - 'null'
          format: date
          example: '2024-03-20'
          description: 'If not provided, the issue date will be the current date.'
        dueDate:
          type:
            - string
            - 'null'
          format: date
          example: '2024-04-20'
          description: 'If not provided, the due date will be 1 month from the issue date.'
        note:
          type:
            - string
            - 'null'
          example: Thank you for your business
        buyerReference:
          type:
            - string
            - 'null'
          example: PO-2024-001
        purchaseOrderReference:
          type:
            - string
            - 'null'
          example: PO-2024-001
        seller:
          oneOf:
            - $ref: '#/components/schemas/Party'
            - type: 'null'
          description: >-
            If not provided, the seller will be the company that is sending the
            invoice.
        buyer:
          $ref: '#/components/schemas/Party'
        paymentMeans:
          type:
            - array
            - 'null'
          items:
            $ref: '#/components/schemas/PaymentMeans'
          description: >-
            Optional payment information. For most invoices, this should be
            provided. For prepaid invoices, this could be omitted.
        paymentTerms:
          type:
            - object
            - 'null'
          properties:
            note:
              type: string
              example: Net 30
          required:
            - note
        lines:
          type: array
          items:
            $ref: '#/components/schemas/Line'
          minItems: 1
        totals:
          oneOf:
            - $ref: '#/components/schemas/Totals'
            - type: 'null'
        vat:
          oneOf:
            - $ref: '#/components/schemas/VatTotals'
            - type: 'null'
        attachments:
          type:
            - array
            - 'null'
          items:
            $ref: '#/components/schemas/Attachment'
          description: Optional attachments to the invoice
      required:
        - invoiceNumber
        - buyer
        - lines
      title: Invoice to send
      description: Invoice to send to a recipient
...
    SendCreditNote:
      type: object
      properties:
        creditNoteNumber:
          type: string
          example: CN-2024-001
        issueDate:
          type:
            - string
            - 'null'
          format: date
          example: '2024-03-20'
          description: 'If not provided, the issue date will be the current date.'
        note:
          type:
            - string
            - 'null'
          example: Thank you for your business
        buyerReference:
          type:
            - string
            - 'null'
          example: PO-2024-001
        invoiceReferences:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                minLength: 1
                example: INV-2024-001
                description: The reference to the invoice that is being credited
              issueDate:
                type:
                  - string
                  - 'null'
                format: date
                example: '2024-03-20'
                description: The issue date of the invoice that is being credited
            required:
              - id
          default: []
          description: References to one or more invoices that are being credited
        purchaseOrderReference:
          type:
            - string
            - 'null'
          example: PO-2024-001
        seller:
          oneOf:
            - $ref: '#/components/schemas/Party'
            - type: 'null'
          description: >-
            If not provided, the seller will be the company that is sending the
            credit note.
        buyer:
          $ref: '#/components/schemas/Party'
        paymentMeans:
          type:
            - array
            - 'null'
          items:
            $ref: '#/components/schemas/PaymentMeans'
        paymentTerms:
          type:
            - object
            - 'null'
          properties:
            note:
              type: string
              example: Net 30
          required:
            - note
        lines:
          type: array
          items:
            $ref: '#/components/schemas/Line'
          minItems: 1
        totals:
          oneOf:
            - $ref: '#/components/schemas/Totals'
            - type: 'null'
        vat:
          oneOf:
            - $ref: '#/components/schemas/VatTotals'
            - type: 'null'
        attachments:
          type:
            - array
            - 'null'
          items:
            $ref: '#/components/schemas/Attachment'
          description: Optional attachments to the credit note
        dueDate:
          type:
            - string
            - 'null'
          format: date
          example: '2024-04-20'
          description: 'If not provided, the due date will be 1 month from the issue date.'
      required:
        - creditNoteNumber
        - buyer
        - lines
      title: Credit Note to send
      description: Credit note to send to a recipient
    SendSelfBillingInvoice:
      type: object
      properties:
        invoiceNumber:
          type: string
          example: INV-2024-001
        issueDate:
          type:
            - string
            - 'null'
          format: date
          example: '2024-03-20'
          description: 'If not provided, the issue date will be the current date.'
        dueDate:
          type:
            - string
            - 'null'
          format: date
          example: '2024-04-20'
          description: 'If not provided, the due date will be 1 month from the issue date.'
        note:
          type:
            - string
            - 'null'
          example: Thank you for your business
        buyerReference:
          type:
            - string
            - 'null'
          example: PO-2024-001
        purchaseOrderReference:
          type:
            - string
            - 'null'
          example: PO-2024-001
        seller:
          oneOf:
            - $ref: '#/components/schemas/Party'
            - type: 'null'
          description: >-
            If not provided, the seller will be the company that is sending the
            invoice.
        buyer:
          $ref: '#/components/schemas/Party'
        paymentMeans:
          type:
            - array
            - 'null'
          items:
            $ref: '#/components/schemas/PaymentMeans'
          description: >-
            Optional payment information. For most invoices, this should be
            provided. For prepaid invoices, this could be omitted.
        paymentTerms:
          type:
            - object
            - 'null'
          properties:
            note:
              type: string
              example: Net 30
          required:
            - note
        lines:
          type: array
          items:
            $ref: '#/components/schemas/Line'
          minItems: 1
        totals:
          oneOf:
            - $ref: '#/components/schemas/Totals'
            - type: 'null'
        vat:
          oneOf:
            - $ref: '#/components/schemas/VatTotals'
            - type: 'null'
        attachments:
          type:
            - array
            - 'null'
          items:
            $ref: '#/components/schemas/Attachment'
          description: Optional attachments to the invoice
      required:
        - invoiceNumber
        - buyer
        - lines
      title: Self Billing Invoice to send
      description: Self billing invoice to send to a recipient
    SendSelfBillingCreditNote:
      type: object
      properties:
        creditNoteNumber:
          type: string
          example: CN-2024-001
        issueDate:
          type:
            - string
            - 'null'
          format: date
          example: '2024-03-20'
          description: 'If not provided, the issue date will be the current date.'
        note:
          type:
            - string
            - 'null'
          example: Thank you for your business
        buyerReference:
          type:
            - string
            - 'null'
          example: PO-2024-001
        invoiceReferences:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                minLength: 1
                example: INV-2024-001
                description: The reference to the invoice that is being credited
              issueDate:
                type:
                  - string
                  - 'null'
                format: date
                example: '2024-03-20'
                description: The issue date of the invoice that is being credited
            required:
              - id
          default: []
          description: References to one or more invoices that are being credited
        purchaseOrderReference:
          type:
            - string
            - 'null'
          example: PO-2024-001
        seller:
          oneOf:
            - $ref: '#/components/schemas/Party'
            - type: 'null'
          description: >-
            If not provided, the seller will be the company that is sending the
            credit note.
        buyer:
          $ref: '#/components/schemas/Party'
        paymentMeans:
          type:
            - array
            - 'null'
          items:
            $ref: '#/components/schemas/PaymentMeans'
        paymentTerms:
          type:
            - object
            - 'null'
          properties:
            note:
              type: string
              example: Net 30
          required:
            - note
        lines:
          type: array
          items:
            $ref: '#/components/schemas/Line'
          minItems: 1
        totals:
          oneOf:
            - $ref: '#/components/schemas/Totals'
            - type: 'null'
        vat:
          oneOf:
            - $ref: '#/components/schemas/VatTotals'
            - type: 'null'
        attachments:
          type:
            - array
            - 'null'
          items:
            $ref: '#/components/schemas/Attachment'
          description: Optional attachments to the credit note
        dueDate:
          type:
            - string
            - 'null'
          format: date
          example: '2024-04-20'
          description: 'If not provided, the due date will be 1 month from the issue date.'
      required:
        - creditNoteNumber
        - buyer
        - lines
      title: Self Billing Credit Note to send
      description: Self billing credit note to send to a recipient
    XML:
      type: string
      title: XML
      description: XML document as a string
...
security:
  - httpBasic: []
...
paths:
  '/api/peppol/{companyId}/sendDocument':
    post:
      summary: Send Document
      operationId: sendDocument
      tags:
        - Sending
      description: Send a document to a customer
      parameters:
        - schema:
            type: string
          in: path
          name: companyId
          required: true
      responses:
        '200':
          description: Successfully sent document
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  sentOverPeppol:
                    type: boolean
                    description: Whether the document was sent over Peppol
                  sentOverEmail:
                    type: boolean
                    description: Whether the document was sent over email
                  emailRecipients:
                    type: array
                    description: The email addresses that the document was sent to
                  teamId:
                    type: string
                    description: The ID of the team that sent the document
                  companyId:
                    type: string
                    description: The ID of the company that sent the document
                  id:
                    type: string
                    description: The ID of the transmitted document
        '400':
          description: Invalid document data provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  errors:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                recipient:
                  type: string
                  description: >-
                    The Peppol address of the recipient of the document. If no
                    identifier is provided, 0208 (Belgian Enterprise Number) is
                    assumed.
                  example: '0208:987654321'
                email:
                  $ref: '#/components/schemas/Email'
                documentType:
                  type: string
                  enum:
                    - invoice
                    - creditNote
                    - selfBillingInvoice
                    - selfBillingCreditNote
                    - xml
                  description: The type of document.
                  example: invoice
                document:
                  anyOf:
                    - $ref: '#/components/schemas/SendInvoice'
                    - $ref: '#/components/schemas/SendCreditNote'
                    - $ref: '#/components/schemas/SendSelfBillingInvoice'
                    - $ref: '#/components/schemas/SendSelfBillingCreditNote'
                    - $ref: '#/components/schemas/XML'
                doctypeId:
                  type: string
                  description: >-
                    The document type identifier. Not required, only used when
                    documentType is "xml".
                  example: >-
                    urn:oasis:names:specification:ubl:schema:xsd:Invoice-2::Invoice##urn:cen.eu:en16931:2017#compliant#urn:fdc:peppol.eu:2017:poacc:billing:3.0::2.1
              required:
                - recipient
                - documentType
                - document
...